test: move all tests to tests/
