+++ jest.config.js
@@ -1,6 +1,6 @@
 /** @type {import('jest').Config} */
 module.exports = {
-  roots: ["<rootDir>/src"],
+  roots: ["<rootDir>/tests"],
   coverageDirectory: "coverage",
   collectCoverageFrom: ["<rootDir>/src/**"],
   testEnvironment: "node",
