+++ tests/main/middlewares/body-parser.test.js
@@ -1,6 +1,6 @@
 const request = require("supertest");
 
-const { app } = require("../config/app");
+const { app } = require("../../../src/main/config/app");
 
 describe("bodyParser middleware", () => {
   let requestHeaders;
